Added 64 bit support. Dubrovnik now links to a 64 bit Mono framework in /Library/Frameworks/Mono64.framework

+++ Framework/XCode/Wrappers/DBWrappedData.m
@@ -66,7 +66,7 @@
 	return(_dataBytes);
 }
 
-- (unsigned)length {
+- (uintptr_t)length {
 	return(_dataLength);
 }
 
@@ -89,7 +89,7 @@
 	memcpy(buffer, _dataBytes, _dataLength);
 }
 
-- (void)getBytes:(void *)buffer length:(unsigned)length {
+- (void)getBytes:(void *)buffer length:(NSUInteger)length {
 	memcpy(buffer, _dataBytes, MIN(_dataLength, length));
 }
